refactor: separate engine heartbeat instances
